Move the pretty printing function from generate_java_generic_type to java_utils.
