+++ unit/java_bytecode/java_utils_test.cpp
@@ -341,3 +341,30 @@ SCENARIO("find_closing_semi_colon_for_reference_type", "[core][java_util_test]")
     REQUIRE(find_closing_semi_colon_for_reference_type(descriptor, 10) == 19);
   }
 }
+
+SCENARIO("Test pretty printing auxiliary function", "[core][java_util_test]")
+{
+  using std::map;
+  using std::string;
+
+  WHEN("We have a series of cbmc internal java types")
+  {
+    // NOLINTNEXTLINE
+    const map<string, string> types{
+      // map<Input, Output>
+      {"java::java.lang.Integer", "Integer"},
+      {"java::CustomClass", "CustomClass"},
+      {"java.lang.String", "String"},
+      {"Hashmap", "Hashmap"},
+      // We shouldn't prune types not imported in default import
+      {"java.util.HashSet", "java.util.HashSet"}};
+
+    THEN("We need to make sure that the types get pruned correctly.")
+    {
+      for(const auto &pair : types)
+      {
+        REQUIRE(pretty_print_java_type(pair.first) == pair.second);
+      }
+    }
+  }
+}
